Fixed bug allowing errors to persist to next page; created BudgetBody and BudgetCopy

+++ src/components/budget/budgetItem/selectBudgetItem.tsx
@@ -1,49 +1,25 @@
 import { useState, useContext } from 'react';
-import { createBudgetItem, deleteBudgetItem } from '../../../misc/apiCalls';
-import { compileBudgetCategoryNames, findCategoryID } from '../../../misc/miscFunctions';
+import { deleteBudgetItem } from '../../../misc/apiCalls';
 import { SelectBudgetItemProps } from './budgetItemTypes';
-import { BudgetContext, BudgetPeriodContext, CategoriesContext } from '../../../misc/context';
+import { BudgetContext } from '../../../misc/context';
 
 export default function SelectBudgetItem ({ 
     budgetItem, 
 }: SelectBudgetItemProps): JSX.Element {
     
-    const { period } = useContext(BudgetPeriodContext)
     const { setBudgetUpdate } = useContext(BudgetContext)
-    const { categories } = useContext(CategoriesContext)
-    const [ choice, setChoice ] = useState(budgetItem.category);
-    const options = compileBudgetCategoryNames(categories);
+    const [ choice ] = useState(budgetItem.category);
     
     function handleChange (event: React.ChangeEvent<HTMLSelectElement> ) {
         const { value } = event.target;
-        if(value) {
-            if(value === 'Delete') {
-                deleteSelectedBudgetItem();
-            }
-            else if(options.includes(value)) {
-                createNewBudgetItem(value);
-                deleteSelectedBudgetItem();
-            }
+        if(value && value === "Delete") {
+            deleteSelectedBudgetItem();
             setBudgetUpdate(true);
         }
     }    
     
     async function deleteSelectedBudgetItem (): Promise<void> {
         await deleteBudgetItem(budgetItem.id);
-    }
-
-    async function createNewBudgetItem (value: string): Promise<void> {
-        setChoice(value);
-        const categoryId = findCategoryID(value, categories);
-        if (categoryId) {
-            const newFields = {
-                category: categoryId,
-                amount: budgetItem.budget_amount,
-                month: period.month,
-                year: period.year
-            }
-            await createBudgetItem(newFields);  
-        }    
     }
 
     return (
@@ -53,11 +29,8 @@ export default function SelectBudgetItem ({
             name='category'
             onChange={ handleChange }
             data-cy={ `budget-item-category-${choice.toLowerCase()}` }>
-            {
-                options.map((option: string, ndx: number) => (
-                    <option key={ ndx } value={ option }>{ option }</option>
-                ))
-            }
+            <option key={ budgetItem.category } value={ budgetItem.category }>{ budgetItem.category }</option>
+            <option key={ `${budgetItem.category}-Delete` } value="Delete">Delete</option>
         </select>
     )
 }
